resolve all comment from Huy
